+++ package/CHANGELOG.md
@@ -1,3 +1,10 @@
+## Next
+
+### What's New
+
+- add `build.runtime` option
+- add `build.bundle` option
+
 ## 0.3.0 (2026-05-09)
 
 ### What's New